feat: added free mode swiper for mobile devices by default
