+++ package.json
@@ -30,6 +30,7 @@
   },
   "dependencies": {
     "@radix-ui/primitive": "^1.0.1",
+    "@reactuses/core": "^4.0.10",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
