+++ src/pages/Home/Home.tsx
@@ -11,8 +11,6 @@ import { Heading } from '@components/Heading/Heading';
 import { Services } from './components/Services/Services';
 
 import { TeamBlock } from './components/TeamBlock/TeamBlock';
-import { Icon } from '@components/Icon/Icon';
-
 
 export function Home() {
   return (
