+++ src/pages/Home/components/TeamBlock/TeamBlock.tsx
@@ -3,10 +3,9 @@ import { Paragraph } from '@components/Paragraph/Paragraph';
 import React from 'react';
 
 import { Slider } from '@components/Slider/Slider';
-import { TEAM_PERSONS } from '../../../../constants';
+import { SWIPER_PROPS_TEAM, TEAM_PERSONS } from '../../../../constants';
 import { SwiperSlide } from 'swiper/react';
 import { TeamPerson } from '@pages/Team/components/TeamPerson/TeamPerson';
-import { SWIPER_PROPS_TEAM } from './SWIPER_PROPS_TEAM';
 
 interface TeamBlockProps {}
 
@@ -21,7 +20,7 @@ export const TeamBlock: React.FC<TeamBlockProps> = () => {
         які надає “UnimayMedia”:
       </Paragraph>
 
-      <Slider swiperProps={SWIPER_PROPS_TEAM}>
+      <Slider swiperProps={SWIPER_PROPS_TEAM} enableFreeMode={false}>
         {TEAM_PERSONS.map((person) => (
           <SwiperSlide key={person.name}>
             <TeamPerson person={person} />
